Refactor test case for broader response verification Updated test cases to include broader response verification. Expanded the assertions to ensure the HTTP status code, response format, content, and pagination details are correctly validated.
