--- src/test/java/tests/AssignmentManagementServiceTest.java
@@ -1,4 +1,0 @@
-package tests;
-
-public class AssignmentManagementServiceTest {
-}
