+++ src/test/java/tests/auth/InstitutionAuthServiceTest.java
@@ -1,4 +1,4 @@
-package tests;
+package tests.auth;
 
 import endpoints.InstitutionAuthEndpoints;
 import io.restassured.response.Response;
@@ -6,6 +6,9 @@ import org.testng.annotations.Test;
 import payload.AdminCredentials;
 import payload.Token;
 import utility.ConfigurationReader;
+
+import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.notNullValue;
 
 public class InstitutionAuthServiceTest {
     Token token = new Token();
@@ -17,8 +20,14 @@ public class InstitutionAuthServiceTest {
         adminCredentials.setPassword(ConfigurationReader.getProperty("institution1Password"));
         Response response = InstitutionAuthEndpoints.getAdminToken(adminCredentials);
         response.then()
-                .log().body()
-                .statusCode(200);
+                .statusCode(200)
+                .contentType("application/json")
+                .body("time", notNullValue())
+                .body("httpStatus", equalTo("OK"))
+                .body("isSuccess", equalTo(true))
+                .body("response.accessToken", notNullValue())
+                .body("response.accessTokenExpiresAt", notNullValue())
+                .body("response.refreshToken", notNullValue());
         token.setAccessToken(response.jsonPath().getString("response.accessToken"));
         token.setRefreshToken(response.jsonPath().getString("response.refreshToken"));
 
@@ -29,17 +38,26 @@ public class InstitutionAuthServiceTest {
         adminCredentials.setPassword("1234");
         Response response = InstitutionAuthEndpoints.getAdminToken(adminCredentials);
         response.then()
-                .log().body()
-                .statusCode(401);
+                .statusCode(401)
+                .contentType("application/json")
+                .body("time",notNullValue())
+                .body("httpStatus", equalTo("UNAUTHORIZED"))
+                .body("header",equalTo("AUTH ERROR"))
+                .body("isSuccess",equalTo(false));
     }
 
     @Test(priority = 2)
     public void adminTokenRefresh() {
         Response response = InstitutionAuthEndpoints.adminTokenRefresh(token.getRefreshToken());
-        response
-                .then()
-                .log().body()
-                .statusCode(200);
+        response.then()
+                .statusCode(200)
+                .contentType("application/json")
+                .body("time", notNullValue())
+                .body("httpStatus", equalTo("OK"))
+                .body("isSuccess", equalTo(true))
+                .body("response.accessToken", notNullValue())
+                .body("response.accessTokenExpiresAt", notNullValue())
+                .body("response.refreshToken", notNullValue());
         token.setAccessToken(response.jsonPath().getString("response.accessToken"));
         token.setRefreshToken(response.jsonPath().getString("response.refreshToken"));
     }
@@ -47,10 +65,12 @@ public class InstitutionAuthServiceTest {
     @Test(priority = 3)
     public void adminInvalidateToken() {
         Response response=InstitutionAuthEndpoints.adminInvalidateToken(token.getAccessToken(),token.getRefreshToken());
-        response
-                .then()
-                .log().body()
-                .statusCode(200);
+        response.then()
+                .statusCode(200)
+                .contentType("application/json")
+                .body("time", notNullValue())
+                .body("httpStatus", equalTo("OK"))
+                .body("isSuccess", equalTo(true));
     }
 
 }
