+++ src/test/java/tests/institution/InstitutionAssignmentManagementService.java
@@ -1,0 +1,4 @@
+package tests.institution;
+
+public class InstitutionAssignmentManagementService {
+}
