+++ src/test/java/tests/user/UserAssignmentManagementServiceTest.java
@@ -1,0 +1,4 @@
+package tests.user;
+
+public class UserAssignmentManagementServiceTest {
+}
