+++ src/test/java/tests/user/UserLocationManagementServiceTest.java
@@ -1,4 +1,4 @@
-package tests;
+package tests.user;
 
 public class UserLocationManagementServiceTest {
 }
