+++ src/test/java/tests/user/UserSelfManagementServiceTest.java
@@ -1,4 +1,4 @@
-package tests;
+package tests.user;
 
 public class UserSelfManagementServiceTest{
 }
